Portar la autocalibracion a IDAC con signo, sacar el limitador y borrar el legacy
Cuatro cosas, todas pedidas por Elias.

1. polarity_reg se escribe. Cada IDAC8 tiene una entrada ipolarity cableada a un
   bit del Control Register del TopDesign, y NINGUN archivo del firmware lo
   escribia: los cuatro quedaban en el reset y la referencia solo podia alejarse
   de Vref para un lado. Ahora psoc_hw_idac_apply_polarity() fija el signo antes
   de escribir la magnitud (ese orden importa: al reves, un cambio de signo pasa
   un instante por la magnitud nueva con el signo viejo, que es el doble del
   escalon pedido). El codigo pasa a ser CON SIGNO, -255..+255: el 0 es
   exactamente Vref, que es el unico punto desde el que se puede corregir para
   los dos lados, y la autoridad queda en +-478 mV.

   En EEPROM la magnitud sigue en el byte por etapa que ya existia y los cuatro
   signos entran en un byte que estaba libre en la misma fila, asi que no cambia
   ni el tamano de la fila ni el CRC.

2. Fuera el limitador de un codigo por muestra. No era una proteccion: era un
   bug. Con CAL_PI_MAX_DAC_STEP_PER_SAMPLE en 1 el PI nunca aplicaba su propia
   ley de control, la dominaba el limitador, y una etapa que necesitaba moverse
   84 codigos tardaba 84 muestras como piso mientras la integral seguia
   cargando. Ahora el default es 0 = sin limite. La macro queda por si alguna
   etapa alguna vez necesita limitar la pendiente de verdad.

3. Escalas reales de esta placa, todo en enteros. Las constantes eran las de un
   VDAC8 de 16 mV por LSB. El bloque es un IDAC8 que con la resistencia a Vref
   forma un DAC de tension hecho a mano: 125 nA sobre 15 kOhm son 1875 uV por
   codigo. Se agrega un bloque de conversiones con nombre porque confundir los
   dominios es el error clasico: 127 codigos de IDAC no son 127 counts de ADC ni
   127 uV. Las conversiones van en int64 y sin precalcular el cociente: son
   98,304 counts por codigo y redondear eso a 98 mete 0,3 % de error sistematico
   en todo el lazo.

   Las ganancias fisicas referencia->tap salen de la matriz D2 medida el
   2026-09-02, no de un modelo: 31 / 101 / 411 / 651 (x1000) para PGA / BP /
   ADDER / LP.

4. El PGA se puede calibrar. La etapa 0 es la unica sin ganancia fija, porque su
   tap es la salida del PGA de entrada y escala con la ganancia que este puesta:
   medido 57,7 / 108,7 / 220,2 / 448,7 uV por codigo para 1x / 2x / 4x / 8x.
   Antes el PI usaba la de 1x para todas y a 8x pedia ocho veces el esfuerzo
   necesario. Que esa etapa tenga la ganancia mas chica es justamente lo que la
   hace calibrable con precision: 57,7 uV por codigo es el paso mas fino de las
   cuatro.

Legacy borrado: el servo lento y la busqueda binaria. Queda el PI como unico
algoritmo, sin selector ni ramas apagadas. El servo se saco resolviendo sus
#if CAL_ALGO_SERVO_ENABLE como los resolvia el compilador (la macro valia 0),
que es una transformacion que por construccion no cambia lo que compila; su
codigo estaba entreverado con funciones vivas y borrar rangos de lineas se
llevaba puesto codigo que el resto del firmware usa.

program_psoc.ps1: se vuelve a grabar solo las filas que el HEX ocupa (272 con
este firmware) en vez de las 4x256. La version que grababa las 1024 verificaba
todo en 0 OK y dejaba el chip SIN ARRANCAR, medido el 2026-09-02; el flujo
validado contra hardware el 2026-07-07 grababa solo las filas ocupadas. Queda
-AllRows para volver al comportamiento viejo a proposito.

Verificado: compila limpio, sin warnings, 67488 B de flash y 17400 B de SRAM.
La aritmetica de conversion entre dominios se verifico aparte replicandola en
enteros (cuidado: C trunca hacia cero y Python redondea hacia abajo).

NO verificado en la placa: el KitProg se desconecto y no volvio en toda la
noche, asi que no se pudo grabar el PSoC. Falta confirmar en banco que el bit
en 1 de polarity_reg sea el sumidero; si sale al reves, alcanza con dar vuelta
PSOC_IDAC_POLARITY_NEGATIVE_BIT.

+++ AcondicionamientoAnalogicoTest/AcondicionamientoAnalogico.cydsn/psoc_selftest.h
@@ -197,7 +197,7 @@ static void st_report_identity(void)
  * slots, y eso no es una falla. */
 static void st_report_eeprom(void)
 {
-    uint8 dac[PSOC_NV_CAL_STAGES];
+    int16 dac[PSOC_NV_CAL_STAGES];   /* codigos CON SIGNO: 0 = referencia en Vref */
     uint8 slot;
     uint8 valid = 0u;
     uint16 mask = 0u;
